[NFC][MHLO] move util funcs to MhloLegalizeUtils.h/cpp
See RFC: https://github.com/llvm/torch-mlir/issues/999
