+++ lib/Conversion/TorchToMhlo/MhloLegalizeUtils.h
@@ -19,6 +19,11 @@
 
 namespace mlir {
 namespace mhlo {
+#ifdef TORCH_MLIR_ENABLE_MHLO_TRUNC_DIMSIZE_TO_I32
+static constexpr size_t kMhloDimSizeBits = 32;
+#else
+static constexpr size_t kMhloDimSizeBits = 64;
+#endif
 
 using mlir::ConversionPatternRewriter;
 
@@ -56,6 +61,23 @@ LogicalResult torchAlphaToMhloTensor(ConversionPatternRewriter &rewriter,
 
 Value promoteAndBroadcast(ConversionPatternRewriter &rewriter, Value input,
                           TensorType outType);
+
+SmallVector<size_t> toPositiveDims(ArrayRef<int64_t> dims, int64_t rank);
+
+// Get the dimension sizes of the input tensor, given the dimension axes
+FailureOr<SmallVector<Value, 4>> getDimSizesOfTensor(PatternRewriter &rewriter,
+                                                     Operation *op, Value value,
+                                                     ArrayRef<int64_t> inpDims);
+
+// Get the dimension sizes of the input tensor
+FailureOr<SmallVector<Value, 4>>
+getDimSizesOfTensor(PatternRewriter &rewriter, Operation *op, Value value);
+
+// Get a tensor that unsqueezed the specified dimensions of the input tensor
+FailureOr<Value> unsqueezeTensor(PatternRewriter &rewriter, Operation *op,
+                                 Value tensor,
+                                 ArrayRef<int64_t> inputUnsqzDims);
+
 } // namespace mhlo
 } // namespace mlir
